# swagger.json
{
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This package provides a REST API to the Ignition Fuel server.",
    "title": "Ignition Fuel Server REST API",
    "contact": {
      "email": "info@openrobotics.org"
    },
    "license": {
      "name": "Apache 2.0"
    },
    "version": "0.1.0"
  },
  "host": "fuel.ignitionrobotics.org",
  "basePath": "/1.0",
  "paths": {
...
    "/subt/logfiles/{id}": {
      "get": {
        "description": "Updates a log file submission (eg. for scoring)",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "logfiles"
        ],
        "summary": "Update a log file",
        "operationId": "getLogFile",
        "responses": {
          "200": {
            "$ref": "#/responses/LogFile"
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      },
      "delete": {
        "description": "Deletes a log file submission",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "logfiles"
        ],
        "summary": "Deletes a log file",
        "operationId": "deleteLogFile",
        "responses": {
          "200": {
            "$ref": "#/responses/LogFile"
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      },
      "patch": {
        "description": "Updates a log file submission (eg. for scoring)",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "logfiles"
        ],
        "summary": "Update a log file",
        "operationId": "updateLogFile",
        "responses": {
          "200": {
            "$ref": "#/responses/LogFile"
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrMsg": {
      "description": "ErrMsg is serialized as JSON, and returned if the request does not succeed\nTODO: consider making ErrMsg an 'error'",
      "type": "object",
      "properties": {
        "errcode": {
          "description": "Internal error code.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrCode"
        },
        "errid": {
          "description": "Generated ID for easy tracking in server logs",
          "type": "string",
          "x-go-name": "ErrID"
        },
        "extra": {
          "description": "Extra information/arguments associated to Error message.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Extra"
        },
        "msg": {
          "description": "Error message.",
          "type": "string",
          "x-go-name": "Msg"
        },
        "remote-address": {
          "description": "Associated request remote address, if applicable",
          "type": "string",
          "x-go-name": "RemoteAddress"
        },
        "route": {
          "description": "Associated request Route, if applicable",
          "type": "string",
          "x-go-name": "Route"
        },
        "user-agent": {
          "description": "Associated request User-Agent, if applicable",
          "type": "string",
          "x-go-name": "UserAgent"
        }
      },
      "x-go-package": "gitlab.com/ignitionrobotics/web/ign-go"
    },
...
  },
  "responses": {
...
    "fuelError": {
      "description": "Ign Fuel error serialized as JSON",
      "schema": {
        "$ref": "#/definitions/ErrMsg"
      }
    },
...
  }
}